VMO-7198 extract resource helpers from resource.ts (for easier export from the lib)
