+++ src/components/interaction-designer/resource-editors/PerModeResourceEditorRow.vue
@@ -36,7 +36,7 @@ import {
   IFlow,
   IResource,
 } from '@floip/flow-runner'
-import {discoverContentTypesFor} from '@/store/flow/resource'
+import {discoverContentTypesFor} from '@/store/flow/utils/resourceHelpers'
 
 const flowVuexNamespace = namespace('flow')
 const builderVuexNamespace = namespace('builder')
