+++ src/lib.ts
@@ -40,6 +40,9 @@ export * from '@/components/interaction-designer/toolbar'
 export * from '@/store/flow/block-types'
 export * from '@/store'
 export * from '@/router/helpers'
+export * from '@/store/flow/utils/resourceHelpers'
+export * from '@/store/flow/utils/importHelpers'
+export * from '@/store/flow/utils/vuexBlockAndFlowHelpers'
 export {default as flowBuilderStore} from '@/store'
 
 // We may need those validation helpers to customize the validation in the consumer side
